Re-architecture the code so that objects have better abstraction.
Up until now, object such as LogicalSwitch, Dlr, etc. used to have
no knowledge about how to convert names to object-ids. That conversion
was done by Nsx class. It was OK because most of the operations
supported were 'GET' or 'POST'. Now I need to support 'PUT' operation
to update the object and turned out that previous architecture was not
a right approach.

Better architecture would be to have objects know both the name
and the corresponding object-id. When PUT is needed, then update the
object and Nsx class simply calls 'PUT' with the object.

Downside of this architecture I see is whenever name conversion is
needed Nsx object is needed, so those classes end up having an Nsx
object as a parameter, which is somewhat messy.

That being said, I think new architecture would be more self-contained
and object oriented.